Extract out the caching
