+++ b2luigi/batch/processes/__init__.py
@@ -4,8 +4,6 @@ import time
 import abc
 import enum
 
-from cachetools import TTLCache
-
 import luigi
 import luigi.scheduler
 
@@ -17,29 +15,6 @@ class JobStatus(enum.Enum):
     successful = "successful"
     aborted = "aborted"
     idle = "idle"
-
-
-class BatchJobStatusCache(abc.ABC, TTLCache):
-
-    def __init__(self):
-        super(BatchJobStatusCache, self).__init__(maxsize=1000, ttl=20)
-
-    @abc.abstractmethod
-    def _ask_for_job_status(self, job_id=None):
-        pass
-
-    def __missing__(self, job_id):
-        # First, ask for all jobs
-        self._ask_for_job_status(job_id=None)
-        if job_id in self:
-            return self[job_id]
-
-        # Then, ask specifically for this job
-        self._ask_for_job_status(job_id=job_id)
-        if job_id in self:
-            return self[job_id]
-
-        raise KeyError
 
 
 class BatchProcess:
